Adding project files
This is where I have all the code and datafiles already created but just never uploaded them to Git. This project currently works, but I want to add more data to the datafiles and add more to the code to try and better predict a finishing position at any given track.